requirements.txt
fastapi
uvicorn
diffusers
torch
Pillow
gradio
requests
transformers 